chore(ci): lint .buildkite/scripts + pay down surfaced debt
.buildkite/scripts/*.ts were typechecked and tested but never linted
(the root-scripts lint ran eslint only in scripts/). New
scripts/lint-buildkite.ts drives the ESLint API with cwd = repo root
(flat-config patterns resolve against cwd, so a CLI invocation from
the scripts workspace ignores files outside it), plus a tsconfig shim
in .buildkite/ so typescript-eslint's project service can discover
the files (the isolated linker hoists no @types to the root).

Fixing the surfaced debt properly, not suppressing it:
- select-image-targets.ts split into three scannable modules
  (selection orchestration / lockfile+manifest+patch attribution /
  workspace-graph closure walking), with the selection walk decomposed
  into per-mechanism helpers; both new files registered in
  GLOBAL_IMAGE_INPUTS so selector changes still rebuild everything.
- validate-pipeline.ts generic helpers extracted to
  validate-pipeline-lib.ts; nullable-boolean conditionals made
  explicit; behavior byte-identical.
- assorted mechanical fixes (unsafe any at the lockfile package walk,
  string-index concat, useless escapes, node:fs → Bun.file).

One scoped rule adjustment, documented in scripts/eslint.config.ts:
custom-rules/no-type-guards is off for .buildkite/scripts because
those scripts run dependency-free via bun --no-install BEFORE any
workspace install — Zod cannot exist there.
